일단, 기본적인 구조 리팩토링은 마쳤음. Now we need to write logic for downloading & storing logic to local mongo db.
